Added Shame poem.
What's it about? It's about the changes in life one goes through, it's
about how looking back on things we can see that we were part or were
the problem. And how reflecting on this sometimes we see a change for
the better in ourselves. But we still have to live with the regret and
shame of our actions before.