Small narration about waking up.
Found a swp file of something I never finished, decided to pick it up
and give it a bit of an ending. I don't think it's what I intended
originally as the file was called relaxed and I think I wanted to write
about how relaxed and good I felt one of the mornings last week. But
it's hard to capture that in writing.